Removed some cruft and got hugo working with build
So now webpack build (so with cachebusting) will write a manifest to
the data directory in json and if it's present then hugo will use it
to output the correct filenames for assets. If it's not there then
hugo will assume we're using the webpack-dev-server so we aren't
having to rebuild all the pages each time we change the css or js
and so we get css / js hot reloading.